Refactor handleStream to LLM Classes
