+++ server/utils/AiProviders/localAi/index.js
@@ -1,4 +1,5 @@
 const { chatPrompt } = require("../../chats");
+const { handleDefaultStreamResponse } = require("../../chats/stream");
 
 class LocalAiLLM {
   constructor(embedder = null, modelPreference = null) {
@@ -172,6 +173,10 @@ class LocalAiLLM {
       { responseType: "stream" }
     );
     return streamRequest;
+  }
+
+  handleStream(response, stream, responseProps) {
+    return handleDefaultStreamResponse(response, stream, responseProps);
   }
 
   // Simple wrapper for dynamic embedder & normalize interface for all LLM implementations
