+++ server/utils/chats/stream.js
@@ -156,7 +156,7 @@ async function streamChatWithWorkspace(
     const stream = await LLMConnector.streamGetChatCompletion(messages, {
       temperature: workspace?.openAiTemp ?? LLMConnector.defaultTemp,
     });
-    completeText = await handleStreamResponses(response, stream, {
+    completeText = await LLMConnector.handleStream(response, stream, {
       uuid,
       sources,
     });
@@ -214,7 +214,7 @@ async function streamEmptyEmbeddingChat({
       workspace,
       rawHistory
     );
-    completeText = await handleStreamResponses(response, stream, {
+    completeText = await LLMConnector.handleStream(response, stream, {
       uuid,
       sources: [],
     });
@@ -229,300 +229,9 @@ async function streamEmptyEmbeddingChat({
   return;
 }
 
-// TODO: Refactor this implementation
-function handleStreamResponses(response, stream, responseProps) {
+// The default way to handle a stream response. Functions best with OpenAI.
+function handleDefaultStreamResponse(response, stream, responseProps) {
   const { uuid = uuidv4(), sources = [] } = responseProps;
-
-  // Gemini likes to return a stream asyncIterator which will
-  // be a totally different object than other models.
-  if (stream?.type === "geminiStream") {
-    return new Promise(async (resolve) => {
-      let fullText = "";
-      for await (const chunk of stream.stream) {
-        fullText += chunk.text();
-        writeResponseChunk(response, {
-          uuid,
-          sources: [],
-          type: "textResponseChunk",
-          textResponse: chunk.text(),
-          close: false,
-          error: false,
-        });
-      }
-
-      writeResponseChunk(response, {
-        uuid,
-        sources,
-        type: "textResponseChunk",
-        textResponse: "",
-        close: true,
-        error: false,
-      });
-      resolve(fullText);
-    });
-  }
-
-  if (stream?.type === "azureStream") {
-    return new Promise(async (resolve) => {
-      let fullText = "";
-      for await (const event of stream.stream) {
-        for (const choice of event.choices) {
-          const delta = choice.delta?.content;
-          if (!delta) continue;
-          fullText += delta;
-          writeResponseChunk(response, {
-            uuid,
-            sources: [],
-            type: "textResponseChunk",
-            textResponse: delta,
-            close: false,
-            error: false,
-          });
-        }
-      }
-
-      writeResponseChunk(response, {
-        uuid,
-        sources,
-        type: "textResponseChunk",
-        textResponse: "",
-        close: true,
-        error: false,
-      });
-      resolve(fullText);
-    });
-  }
-
-  if (stream.type === "togetherAiStream") {
-    return new Promise((resolve) => {
-      let fullText = "";
-      let chunk = "";
-      stream.stream.data.on("data", (data) => {
-        const lines = data
-          ?.toString()
-          ?.split("\n")
-          .filter((line) => line.trim() !== "");
-
-        for (const line of lines) {
-          let validJSON = false;
-          const message = chunk + line.replace(/^data: /, "");
-
-          if (message !== "[DONE]") {
-            // JSON chunk is incomplete and has not ended yet
-            // so we need to stitch it together. You would think JSON
-            // chunks would only come complete - but they don't!
-            try {
-              JSON.parse(message);
-              validJSON = true;
-            } catch {}
-
-            if (!validJSON) {
-              // It can be possible that the chunk decoding is running away
-              // and the message chunk fails to append due to string length.
-              // In this case abort the chunk and reset so we can continue.
-              // ref: https://github.com/Mintplex-Labs/anything-llm/issues/416
-              try {
-                chunk += message;
-              } catch (e) {
-                console.error(`Chunk appending error`, e);
-                chunk = "";
-              }
-              continue;
-            } else {
-              chunk = "";
-            }
-          }
-
-          if (message == "[DONE]") {
-            writeResponseChunk(response, {
-              uuid,
-              sources,
-              type: "textResponseChunk",
-              textResponse: "",
-              close: true,
-              error: false,
-            });
-            resolve(fullText);
-          } else {
-            let finishReason = null;
-            let token = "";
-            try {
-              const json = JSON.parse(message);
-              token = json?.choices?.[0]?.delta?.content;
-              finishReason = json?.choices?.[0]?.finish_reason || null;
-            } catch {
-              continue;
-            }
-
-            if (token) {
-              fullText += token;
-              writeResponseChunk(response, {
-                uuid,
-                sources: [],
-                type: "textResponseChunk",
-                textResponse: token,
-                close: false,
-                error: false,
-              });
-            }
-
-            if (finishReason !== null) {
-              writeResponseChunk(response, {
-                uuid,
-                sources,
-                type: "textResponseChunk",
-                textResponse: "",
-                close: true,
-                error: false,
-              });
-              resolve(fullText);
-            }
-          }
-        }
-      });
-    });
-  }
-
-  if (stream.type === "huggingFaceStream") {
-    return new Promise((resolve) => {
-      let fullText = "";
-      let chunk = "";
-      stream.stream.data.on("data", (data) => {
-        const lines = data
-          ?.toString()
-          ?.split("\n")
-          .filter((line) => line.trim() !== "");
-
-        for (const line of lines) {
-          let validJSON = false;
-          const message = chunk + line.replace(/^data:/, "");
-          if (message !== "[DONE]") {
-            // JSON chunk is incomplete and has not ended yet
-            // so we need to stitch it together. You would think JSON
-            // chunks would only come complete - but they don't!
-            try {
-              JSON.parse(message);
-              validJSON = true;
-            } catch {
-              console.log("Failed to parse message", message);
-            }
-
-            if (!validJSON) {
-              // It can be possible that the chunk decoding is running away
-              // and the message chunk fails to append due to string length.
-              // In this case abort the chunk and reset so we can continue.
-              // ref: https://github.com/Mintplex-Labs/anything-llm/issues/416
-              try {
-                chunk += message;
-              } catch (e) {
-                console.error(`Chunk appending error`, e);
-                chunk = "";
-              }
-              continue;
-            } else {
-              chunk = "";
-            }
-          }
-
-          if (message == "[DONE]") {
-            writeResponseChunk(response, {
-              uuid,
-              sources,
-              type: "textResponseChunk",
-              textResponse: "",
-              close: true,
-              error: false,
-            });
-            resolve(fullText);
-          } else {
-            let error = null;
-            let finishReason = null;
-            let token = "";
-            try {
-              const json = JSON.parse(message);
-              error = json?.error || null;
-              token = json?.choices?.[0]?.delta?.content;
-              finishReason = json?.choices?.[0]?.finish_reason || null;
-            } catch {
-              continue;
-            }
-
-            if (!!error) {
-              writeResponseChunk(response, {
-                uuid,
-                sources: [],
-                type: "textResponseChunk",
-                textResponse: null,
-                close: true,
-                error,
-              });
-              resolve("");
-              return;
-            }
-
-            if (token) {
-              fullText += token;
-              writeResponseChunk(response, {
-                uuid,
-                sources: [],
-                type: "textResponseChunk",
-                textResponse: token,
-                close: false,
-                error: false,
-              });
-            }
-
-            if (finishReason !== null) {
-              writeResponseChunk(response, {
-                uuid,
-                sources,
-                type: "textResponseChunk",
-                textResponse: "",
-                close: true,
-                error: false,
-              });
-              resolve(fullText);
-            }
-          }
-        }
-      });
-    });
-  }
-
-  // If stream is not a regular OpenAI Stream (like if using native model, Ollama, or most LangChain interfaces)
-  // we can just iterate the stream content instead.
-  if (!stream.hasOwnProperty("data")) {
-    return new Promise(async (resolve) => {
-      let fullText = "";
-      for await (const chunk of stream) {
-        if (chunk === undefined)
-          throw new Error(
-            "Stream returned undefined chunk. Aborting reply - check model provider logs."
-          );
-
-        const content = chunk.hasOwnProperty("content") ? chunk.content : chunk;
-        fullText += content;
-        writeResponseChunk(response, {
-          uuid,
-          sources: [],
-          type: "textResponseChunk",
-          textResponse: content,
-          close: false,
-          error: false,
-        });
-      }
-
-      writeResponseChunk(response, {
-        uuid,
-        sources,
-        type: "textResponseChunk",
-        textResponse: "",
-        close: true,
-        error: false,
-      });
-      resolve(fullText);
-    });
-  }
 
   return new Promise((resolve) => {
     let fullText = "";
@@ -615,5 +324,5 @@ module.exports = {
   VALID_CHAT_MODE,
   streamChatWithWorkspace,
   writeResponseChunk,
-  handleStreamResponses,
+  handleDefaultStreamResponse,
 };
